get all leagues joined by working
