--- StockSeer/src/view/model/AvailableStockModel.java
@@ -1,32 +1,0 @@
-package view.model;
-
-import javax.swing.ListModel;
-import javax.swing.event.ListDataListener;
-
-public class AvailableStockModel implements ListModel {
-
-    public void addElement(String temp) {
-
-    }
-
-    @Override
-    public int getSize() {
-        return 0;
-    }
-
-    @Override
-    public Object getElementAt(int index) {
-        return null;
-    }
-
-    @Override
-    public void addListDataListener(ListDataListener l) {
-
-    }
-
-    @Override
-    public void removeListDataListener(ListDataListener l) {
-
-    }
-
-}
